backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
